Extract useRequestPipeline hook from EndpointDetail
EndpointDetail was carrying nine useState hooks and the full execute
/abort/reset choreography alongside its display logic. Move the
request-pipeline state and actions into a dedicated hook so the
component can stay focused on rendering and mode toggling.
 - New hook owns parameterValues, requestBodyValue, selectedContentType,
   customQueryParams, customHeaders, requestData, responseData,
   isLoading, and the AbortController lifecycle.
 - Derived flags (requiredParametersMissing, requiredBodyMissing) move
   with the state they derive from; server-URL validity stays in
   EndpointDetail since it depends on the serverUrl prop.
 - Actions exposed as execute / abort / reset; Done in the UI composes
   reset + setIsInteractive(false).
 - No behavior change; existing tests still pass.
